hooked up profile page and logout function
