+++ client/src/components/Progress/Progress.scss
@@ -1,17 +1,21 @@
+@use "../../styles/partials/mixins" as *;
+@use "../../styles/partials/variables" as *;
+
 .progress {
   background-color: #d8d8d8;
   border-radius: 20px;
   position: relative;
   margin: 15px 0;
-  height: 30px;
+  height: 20px;
   width: 300px;
 }
 
 .progress-done {
-  background: linear-gradient(to left, #f2709c, #ff9472);
-  box-shadow: 0 3px 3px -5px #f2709c, 0 2px 5px #f2709c;
+  background: linear-gradient(to left, #4696f7, #9ac3f5);
+  // box-shadow: 0 2px 2px -5px #76d9fa, 0 2px 3px #388ef8;
   border-radius: 20px;
-  color: #fff;
+  color: rgb(10, 9, 9);
+  @include mobil-labels;
   display: flex;
   align-items: center;
   justify-content: center;
