+++ client/src/pages/ProfilePage/ProfilePage.scss
@@ -1,0 +1,88 @@
+@use "../../styles/partials/mixins" as *;
+@use "../../styles/partials/variables" as *;
+
+.profile {
+  @include main-container;
+  position: relative;
+
+  &__logout {
+    width: 1.5rem;
+    height: 1.5rem;
+    position: absolute;
+    right: 2rem;
+    top: 2rem;
+    opacity: 0.9;
+
+    @include tablet {
+      width: 2rem;
+      height: 2rem;
+    }
+
+    &:hover {
+      cursor: pointer;
+      transform: scale(1.2);
+      opacity: 1;
+    }
+  }
+
+  &__top {
+    width: 100%;
+    @include flexbox(column, flex-start, center);
+  }
+
+  &__greeting {
+    @include mobil-subheader;
+    margin-bottom: 2rem;
+
+    @include tablet {
+      @include subheader;
+    }
+  }
+
+  &__strong {
+    @include mobil-page-header;
+    color: $primary-blue;
+    margin-left: 0.5rem;
+
+    @include tablet {
+      @include page-header;
+    }
+  }
+
+  &__container {
+    width: 6.25rem;
+    height: 6.25rem;
+    border-top-left-radius: 3rem;
+    border: 1px solid $primary-blue;
+    box-shadow: 1px 2px 4px #b8e4f9fe;
+  }
+
+  &__image {
+    object-fit: cover;
+    border-top-left-radius: 3rem;
+  }
+
+  &__bottom {
+    margin-top: 2rem;
+    width: 100%;
+    @include flexbox(column, flex-start, center);
+    @include tablet {
+      margin-top: 4rem;
+    }
+  }
+
+  &__progress-title {
+    @include mobil-subheader;
+    @include tablet {
+      @include subheader;
+    }
+  }
+
+  &__progress {
+    width: 90%;
+
+    @include tablet {
+      width: 80%;
+    }
+  }
+}
